Move mix data into likelihood function
In prior commits, `think-bayes.cookie.likelihood` was a closure over the
variable `think-bayes.cookie.mix`. The `likelihood` function depends on
`mix` but nothing else does, so this commit moved `mix` inside the scope
of `likelihood`.

In addition, in prior commits, `think-bayes.cookie.posterior` was a
closure over the variable `think-bayes.cookie.likelihood`. This commit
changes the `posterior` function to accept a `likelihood` paramater.