Probability of bowl 1 given vanilla cookie
Write code to calculate the probability that a cookie was selected from
"bowl 1" given that the cookie was a vanilla cookie.

This implementation uses the table method for calculating probabilities
using Bayes' law.

As part of this implementation, write general function to calculate the
posterior probabilities given priors, likelihoods and a datum (called
data).